buddybill/views.py
from django_filters.rest_framework import DjangoFilterBackend
from rest_framework import viewsets

from . import permissions
from .models import Group
from .serializers import GroupSerializer


class GroupViewSet(viewsets.ModelViewSet):
    queryset = Group.objects.all()
    serializer_class = GroupSerializer
    permission_classes = (permissions.GroupPermission,)
    filter_backend = [DjangoFilterBackend]
    filterset_fields = {
        "name": ["icontains"],
    }
