requirements.txt
boto3==1.35.2
celery==5.4.0
coverage==7.6.1
dj-database-url==2.2.0
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-allauth==64.2.1
django-anymail==12.0
django-filter==24.3
dj-rest-auth==6.0.0
factory-boy==3.3.1
gunicorn==23.0.0
pillow==10.4.0
pre-commit==3.8.0
pyyaml==6.0.2
python-dotenv==1.0.0
ruff==0.6.6
uritemplate==4.1.1
whitenoise==6.7.0
